Make pre-2.1.0 installation a successful no-op.
curses is part of ruby-core pre-2.1.0, so this gem is not required for
such versions. Previously, installation failed on Ruby 1.9.3 (but was
successful on Ruby 2.0.0).

    https://github.com/ruby/curses/issues/1

Subsequently, `required_ruby_version` was set in the `gemspec` to
clarify that only 2.1.0 and later are supported.

    https://github.com/ruby/curses/commit/4e55ed5c82af33743570a9daa05ef0884a79c06a

But this still means that gems attempting to offer multi-version Ruby
support fail, because of not being able to specifify `gem 'curses'`
within the `gemspec`. Instead, use the referenced resolution within

    https://github.com/guard/listen/issues/83

as inspiration, and make pre-2.1.0 installation result in a no-op. In such
a case, `require 'curses'` is expected to work as usual.