Need to tell whether PDCurses is dll or not
With VC++, we have to tell to linker that a referenced symbol is in DLL if
the symbol is not a function but a variable by using `__declspec(dllimport)`.
`curses.h` of PDCurses provides the way to declare symbols with it, but not
enabled by default.
Therefore, we have to tell it by defining `PDC_DLL_BUILD`.