Use `addnwstr` and `waddnwstr` if they are available
In old implementation, we convert the passed string to terminal_encoding and
pass the result to `addstr` or `waddstr`.
If using PDCurses, in `addstr` and `waddstr` the passed string is converted to
wide char by using `mbstows`.
But `mbstows` requires to call `setlocale` to know the encoding of passed
multibyte string.
Since `setlocale` affects entire process, we cannot call it only for this
purpose.
Then we have to convert the original string to wide char by ourselves and pass
the result to wide string APIs of curses.